Rename: 공지사항 작성 에디터 컴포넌트 이름 변경
